requirements.txt
asyncpg~=0.29.0
fastapi~=0.113.0
pydantic[email]~=2.9.2
pydantic-settings~=2.5.2
python-dotenv~=1.0.1
sqlalchemy~=2.0.22
transliterate~=1.10.2
uvicorn~=0.31.1